Log platform correctly
`log` is a chef resource which is updated at each run.
For users willing to decrease the amount of resources considered at each
run (to monitor the updates done by chef), this is necessary to use
Chef::Log mechanism instead.

In practice it will move the log line from convergence phase (when
resource are converged/executed) to compilation phase (when recipe code
is read)

Change-Id: Ia6599b30b2bcb2ce585f1f15653c2f389fc94047
Signed-off-by: Grégoire Seux <g.seux@criteo.com>